+++ src/Trarizon.Library/TraThrow.cs
@@ -18,6 +18,10 @@ internal static class TraThrow
         => throw new InvalidOperationException("Collection has no element.");
 
     [DoesNotReturn]
+    internal static void NodeBelongsWrong() 
+        => throw new InvalidOperationException("Node does not belong to this collection.");
+
+    [DoesNotReturn]
     internal static void Exception(Exception exception)
         => throw exception;