+++ tests/Trarizon.Test.Run/Program.cs
@@ -26,6 +26,7 @@ using Trarizon.Library.CodeGeneration;
 using Trarizon.Library.Collections;
 using Trarizon.Library.Collections.AllocOpt;
 using Trarizon.Library.Collections.Generic;
+using Trarizon.Library.Collections.Specialized;
 using Trarizon.Library.Numerics;
 using Trarizon.Library.RunTest.Examples;
 using Trarizon.Library.Text;
@@ -34,7 +35,16 @@ using Trarizon.Library.Threading;
 using Trarizon.Library.Wrappers;
 using Trarizon.Test.Run;
 
-Trie<char> chs = ["str", "adb"];
+var list=new ContiguousLinkedList<int>();
+list.AddLast(1);
+list.AddLast(2);
+list.Print();
+list.AddAfter(list.FirstNode, 3);
+list.Print();
+list.AddFirst(4);
+list.Print();
+list.Remove(list.FirstNode.Next);
+list.Print();
 
 class C : IEnumerable<C>
 {